job stat cpu usage precision, add gstreamill cpu usage in stat/gstreamill api

+++ src/job.h
@@ -78,8 +78,8 @@ struct _Job {
     guint64 last_stime; /* last process system time */
     guint64 last_ctime; /* last process cpu time */
     guint64 start_ctime; /* cpu time at process start */
-    gint cpu_average;
-    gint cpu_current;
+    gdouble cpu_average;
+    gdouble cpu_current;
     guint64 memory;
 
     Source *source;
